Fix non-repeatable read issue after after GC pass
Once the snapshot threshold is reached, the in-memory snapshot map is
pruned, therefore, subsequent reads of old snapshots will need to read
from the (disk) snapshot log. An error in the implementation of
materializer_vnode:pvc_internal_read/4 meant that a base value was
always being returned.

The log was actually returning both a base value, and a list of
operations that the caller should apply. We forgot to do that.

This meant that the operations on the key were actually being ignored.
Add a new function to go through all the committed operations and apply
them to create a new snapshot.