Remove disk-log dependency for MaxVC scan during reads
Old version would fetch from disk all the commit records, and
materialize a CLog instance on demand. This change introduces an
in-memory CLog, implemented with an ets table inside logging_vnode.

This way, we don't depend on having a disk-based storage for the reads
to work. Note that the in-memory cache will only exist if the logging is
disabled, otherwise we keep using the old method (probably not ideal?)