Reimplement CLog as a tree
Remove old ets-based version, and reimplement as gb_tree for faster
access. We will refill this log when starting Antidote from the previous
replay log.

In addition, we change the way we find the Maximum VC during a
ReadRequest to pass a cutoff time to the logging_vnode, so we don't copy
the entire CLog back to clocksi_vnode to filter later.