Fix read_index
1. The previous version sometimes would return lists that were not
really appropriate, for example, given that the index keys follow the
format `<grouping>+<index_name>$<foreign_key>$<value>$key`, doing a scan
where the root key is <foreign_key> := user_1, would also yield all keys
for <foreign_key> := [user_10, user_11, ....], as we were not doing
bounds checking correctly. This commit fixes that by always appending
the index separator, `$`, to the root key before scanning.

2. `read_index_range` does not care about the visibility of the index
keys, and just returns a range, where the keys point to the main
versioned storage. It then performs a normal read of those keys, leaving
it to that function to determine the appropiate version to read.
Sometimes, our transaction might not be able to read from an index,
because the entry is too new for our VC. In that case, `read_keys` would
return the base version, `<<>>`. Because the previous version didn't
account for that, it would try to read from `<<>>` as if it were a
returned key from the index, and crash on unexpected input.

The new version purges the base versions resulting from `read_index`.
This is safe, since our grouping and partitioning scheme makes sure that
inserts on index and indexed tables occur atomically and on the same
version. So, even if we skip base values from the index read, we
wouldn't be able to read the appropiate version from the referenced
table anyway, since it would have also return an outdated (or base)
value of the foreign key.