Fix read_objects not updating the transaction
The previous implementation fetched all the keys in advance, and then looped
over the results, updating the transaction in bulk. This is wrong, since it
caused all the original keys to be read as if they were the first key to be
read, not taking into account changes to the `hasRead`, `VCdep` and `VCaggr`
fields of the transaction. This could transactions to observe inconsistent
states (see test/runtime_tests:multiple_read/1).

This commit changes the read_objects logic so the transaction is updated on each
read key, so that each key in the sequence will be read according to a
consistent state.