Implement remote reads using the protocol buffer server
* Each coordinator thread will now hold open sockets to the remote
nodes, and use those to request a remote read operation, instead of
going through the gen_servers

* These sockets should not be reset when the coordinator is returned to
the pool, so we keep them.

* To know the nodes to connect to, we have to wait until we're connected
to the ring to open the sockets. To do that, we won't start the pool
automatically on start up, and do that manually from inter_dc_manager.
To be able to delay, we introduce an intermediate supervisor that we can
command to start the pool manually.