Use separate pool of processes for remote reads
Opening the sockets directly in the coordinators cause overload (too many
sockets open—4 nodes in DC * 1000 coordinator threads). With this change we have
a separate number of processes handling the remote reads using a gen_server.

In the end it doesn't seem like too much of a change compared with the old
implementation of going through globally-registered gen_servers, but on the
other side we should have less overload (as we're using ranch instead of
targetting an individual process on another machine).