Implement database load by force-advancing partition state
Before, we were performing a blind write of N keys with a default value, and
then going through the default 2pc mechanism to commit it. This is ok, but
scales poorly when the number of keys is very big. For millions of keys, it
could take a very long time.

The new approach consists of a) Allowing partitions to return a default value
when reading a key they have never seen, and b) Force-update the partition state
(LastPrep, MostRecentVC and the last commit time for any given key).

By doing both a) and b), we can force the database to be on any state we desire.
Nevertheless, this doesn't stress the database as much as the old method, since
we're not actually storing values in the database, just setting the partition
state by hand.

This procedure is necessary if we're going to move the coordinator to the
client, because the writeset gets too big if we perform a load using the old
mechanism. We implement the new load for the server-side coordinator in order to
perform fair comparions between the two implementations.