Auto-schedule commit dequeue events
Instead of manually requesting a dequeue event after every decide
operation, run an autoschedule operation in the background, configured
by default to run every 1000ms.

When a lot of transactions are comitting, we will flood our process
inbox with `process_queue` messages. Because or dequeue event can
process transactions in batches, it doesn't make sense for it to happen
on every transaction. This way we might get better performance, at the
risk of getting more aborted transactions if enough transactions try to
read new state after a commit, as it will take up to 1 second for
changes to materialize in the storage. We can play with this interval to
see what makes sense.