Use ETS to check for writeset overlap instead of the Commit Queue.
Using an ETS of `bag` type, we can create an inverted index of key =>
tx_id. Checking for writeset overlap now is quite easy, just loop over
our transaction writeset (which is usually small), and check for key
membership on the ETS.

On dequeuing from the commit queue, we can do the same with the
`ets:delete_object/2` operation. We need to remove the keys both when
comitting _and_ aborting the transaction, otherwise we will get fake
aborts.