Move the decide logic out of the vnode
Most of the scalability problems of the 2pc phase comes from the fact
that we serialize client messages through the virtual node. Most of
these messages _need_ to be serialized, since we need some atomicity
guarantees when dealing with the partition state (appending to the
commit queue, updating the version and commit log, etc).

The messages coming as part of the decide part, however, can avoid
routing through the vnode, if we move part of the commit queue logic to
ETS tables. In the old version, the decision would get stored inside the
commit queue state, as `ready_tx` and `discared_tx` fields. This forced
us to get inside the vnode state to be able to modify it. By moving
these two fields to an ETS table, the decide table, we're able to
perform all the logic from outside the vnode process. The clients can
perform a single atomic write against this table using the transaction
id as the key, and the decision as the value, as in `{tx_id, abort} |
`tx_id, ready, VC}`.

In addition to this, we also need to remove all the keys from the
`pending_keys` table. Since this in an inverted index over a transaction
writeset, it's not easy to delete all the keys belonging to a single
transaction. There's `ets:match_delete/2`, but it will have to scan the
entire table to do so. By moving the writeset storage from the commit
queue to an ETS table, we get a tx_id => writeset mapping that we can
use it to quickly delete all the keys belonging to a single transaction.

This leaves us with a commit queue that is barely a wrapper around the
built-in `queue` module, since most of the extra state has been moved
off-heap to ETS tables. This should also help with memory pressure,
since the process doesn't have to store barely anything sans a simple
list of transaction identifiers, which should take up much.