Implement early-abort check outside of clocksi_vnode
Since the data needed to perform the two checks during a prepare
(write-write conflict and stale check) are implemented in terms of ETS
tables, we can move the entire check out of the vnode.

However, due to the parallel nature of this check, combined with the
sequential behaviour of the commit queue, we might get a data race where
two conflicting transactions perform the check in parallel, and both
checks return `valid`, given that the commit queue is empty. They will
then get serialized in the virtual node message inbox. After that, the
first transaction will clear, whereas the second one should abort, but
already thinks it has to commit. This is why we need to double-check the
commit queue even if we already did from the outside.

Even if this check can return false negative, it is still useful to find
early aborts, that is, transaction that would abort no matter what,
either because there is already a conflicting transaction in the queue,
or they are very stale. This saves the penalty of going through the
virtual node and paying the serialization price. Since our main
bottleneck is the virtual node message queue, anything we can do to
reduce the pressure will surely help.