Simplify serializability prepare check
Write-write conflicts are redundant in the presence of read after write
check, since the writeset will always be a subset of the readset
(disallow blind updates).

In the same vein, the stale write check is redundant in the presence of
the stale read. Stale writes are less strict than stale reads, so it's
not needed.