Remove TODO, add commments to CLog
The TODO doesn't make sense, because the abort check specifies that the
clock will never be _smaller_ than the value at VCaggr[i]. If the
smallest value in the tree is larger than that value, it just means that
any element in the tree is potentially valid. A good optimization might
be to check if the largest value in the tree is smaller than VCaggr[i].
In that case, no element in the tree will be valid. Another one is to
always keep in mind that invariant as we walk the tree. As soon as we
step out of bounds, return. But even then, we have to know if we're not
temporarily out of bounds: during the binary search we might step out of
bounds but will eventually stay inside.

In any case, the Largest < VCaggr[i] check is superseded by the wait
until MRVC[i] >= VCaggr[i], which, by definition, if true, it means that
the largest value in the log will also be valid.