Log CLog gc misses
The check is not as straightforward as the one in the version log, but
it's a rough approximation. Even if a GC miss is counted, it doesn't
mean that lack of GC would've fixed the problem, since it still could
happen that the correct clock is not in our log.