Separate per-key and per-partition GC settings
Since the GC mechanism is roughly the same, we used the same settings
for the version and commit log, even if their usage is radically
different.

Keys get a new versions every time they are written, which shouldn't be
too often with a large number of keys and under an uniform distribution
(this could be changed in the future so we may want to fine-tune GC).

For the commit log, however, a new entry is created every time a
transaction commits in a partition. Given this, the amount of versions
being kept should be very different for both.