Delay reads until commit queue has been processed
In addition to waiting until the current MRVC is valid, also check if
the last entry of the queue at READREQUEST time has been decided (if
aborted, wait until removed from the queue, if committed, wait until its
state has been materialized in the partition).

To do so, we have to wait to prune the transaction data until it has
been aborted or materialized in the partition state. Previously, we
removed this state data during dequeue, cleaning space as soon as
possible. There were two problems with this:

1. We were not cleaning up possible RC data on abort (even if not
possible, abort was not covering the case, which could lead to memory
leaks).

2. It is not possible to tell if a transaction has materialized in the
partition state just by looking at the pending data table, as there
could be a time gap between dequeue (when ets:take/2 happens), and the
time the partition state is updated.

Point 2. makes it possible to implement our read until empty queue
without too much overhead. Previously, the only data being kept until
materialization was the pending reads/writes, which is a duplicate bag
indexed by key, not by transaction id. This makes it too expensive to
look up if a specific tx id is no longer in the queue.