Fix maximum read version during first read
Instead of picking the max version allowed during the first read (which
is always the MRVC), look at the prepare time of the last transaction in
the commit queue. This is effectively speculating on the result of the
last prepared transaction. If the transaction commits, we get to read
its result. If the transaction aborts, we will read the values of the
last committed transaction.

This effectively linearizes the reads after the last prepared
transaction at the partition, but before the next one.

We will do this only for the first read, which is the one that fixes the
clock constraints for the rest of the reads. If this change doesn't
diminish the number of read aborts to an acceptable range, we can
implement this for all the reads in a transaction.

We add a new function to the commit log, `get_max_entry_at/2`, which
will return the max entry with a commit time lower or equal than the one
supplied.