Fix maximum allowed version for all reads
If we're waiting for the current LastPrep, we can merge that with the
MRVC wait.

Instead of waiting for the last prepared transaction to be materialized,
wait until the MRVC in the partition goes over that value. This means we
might wait more: if the transaction aborts, we'll wait until another
transaction manages to commit successfully. Hopefully the wait is not
too big.

Once we're done waiting, we still read only UP TO the version we've read
from the queue. Because we do this on every read, we add that threshold
time to our VCaggr, and perform a MaxVC scan as before (add another
partition as a scan constraint).